requirements.txt
flask
gunicorn
pandas
scikit-learn
numpy
joblib
python-dotenv